% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account.phrases}
\alias{account.phrases}
\alias{account.phrases.brandseye.account}
\alias{account.phrases.character}
\alias{account.phrases.list}
\title{List phrases in an account}
\usage{
account.phrases(account, ...)

\method{account.phrases}{brandseye.account}(account, .process = TRUE)

\method{account.phrases}{character}(account)

\method{account.phrases}{list}(accounts)
}
\description{
The returns a \code{data.frame} listing the phrases in an
account, their IDs, the IDs of the brands that the phrase is associated
with, and a flag indicating whether the phrase is inactive or deleted.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: Returns phrase information for a particular account object.

\item \code{character}: Returns account information for an account identified using an account code

\item \code{list}: Given a list of \code{account} objects, this will return
a combined \code{data.frame} for the phrases in all of those accounts.
}}
\examples{
\dontrun{
# Returns account information for an account object
ac <- account("QUIR01BA")
account.phrases(ac)
}
\dontrun{
account.phrases("QUIR01BA")
}
\dontrun{
account.phrases(c("QUIR01BA", "BEAD33AA"))
}
}

