% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mentions.R
\name{account_mentions}
\alias{account_mentions}
\alias{account_mentions.brandseye.account}
\alias{account_mentions.character}
\title{Read mentions from your account}
\usage{
account_mentions(account, ...)

\method{account_mentions}{character}(code, filter, limit = 30, offset = 0,
  include = NULL, authentication = pkg.env$defaultAuthentication)

\method{account_mentions}{brandseye.account}(account, filter)
}
\arguments{
\item{limit}{The maximum number of mentions to be returned}

\item{offset}{Mentions are returned in an order. Offset says how many of the
first mentions should be skipped.}
}
\description{
Read mentions from your account
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{brandseye.account}: 
}}
\examples{
\dontrun{

# Read mentions using the default authentication
account_mentions("QUIR01BA", "published inthelast day")

}
}

