% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_name}
\alias{account_name}
\alias{account_name.brandseye.account}
\alias{account_name.character}
\alias{account_name.factor}
\title{Returns the human readable name of an account}
\usage{
account_name(account)

\method{account_name}{brandseye.account}(account)

\method{account_name}{character}(accounts)

\method{account_name}{factor}(accounts)
}
\description{
Returns the human readable name of an account
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: 

\item \code{character}: 

\item \code{factor}: 
}}
\author{
Rudy Neeser
}

