% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_storage}
\alias{account_storage}
\alias{account_storage.brandseye.account}
\alias{account_storage.character}
\alias{account_storage.factor}
\title{Indicates the underlying storage type of the account.}
\usage{
account_storage(account)

\method{account_storage}{brandseye.account}(account)

\method{account_storage}{character}(accounts)

\method{account_storage}{factor}(accounts)
}
\description{
Indicates the storage type of an account, such as whether it's V4 or the older V3.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: Returns the storage from an \code{\link{account}} object.

\item \code{character}: Returns the storage from a vector of account codes.

\item \code{factor}: Returns the storage of an account given by a factor of the account code.
}}

\author{
Constance Neeser
}
