% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_timezone}
\alias{account_timezone}
\alias{account_timezone.brandseye.account}
\alias{account_timezone.character}
\alias{account_timezone.factor}
\title{Gives the timezone that the account is in.}
\usage{
account_timezone(account)

\method{account_timezone}{brandseye.account}(account)

\method{account_timezone}{character}(accounts)

\method{account_timezone}{factor}(accounts)
}
\description{
Gives the timezone that the account is in.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: Returns the timezone from an \code{\link{account}} object.

\item \code{character}: Returns the timezone from a vector of account codes.

\item \code{factor}: Returns the timezone of an account given by a factor of the account code.
}}

\author{
Constance Neeser
}
