% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\alias{count.brandseye.account}
\alias{count.character}
\title{Count mentions}
\usage{
count(account, ...)

\method{count}{character}(accounts, filter = NULL, groupby = NULL,
  include = NULL, authentication = defaultAuthentication)

\method{count}{brandseye.account}(account, filter = NULL, groupby = NULL,
  include = NULL)
}
\arguments{
\item{account}{An account object to be queried.}

\item{accounts}{A vector of account codes. If this is a single account code, this function
will return a data frame of results just from that account. If it contains multiple accounts,
this will return a data frame containing all the results across accounts, and a column indicating
the account that the particular result is from.}

\item{filter}{A filter string describing the mentions that should be counted by this query
Count aggregate mention information from your BrandsEye account}

\item{groupby}{A list of items that should be grouped by}
}
\description{
\code{count} is used to count mentions in a BrandsEye account matching
matching a particular filter. It's possible to group mentions, order the
results, and to include various other bits of useful information. It's also
possible to perform count operations across multiple accounts.
}
\details{
Filters are described in the api documentation \url{https://api.brandseye.com/docs}
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{brandseye.account}: 
}}
\examples{
count("QUIR01BA", "published inthelast month") # Uses default authentication,
                                               # if that has been set up.
count("QUIR01BA", "published inthelast month",
      authentication = authentication(key = "<my key>"))
# Return results for multiple accounts
count(c("QUIR01BA", "BEAD33AA"), "published inthelast month")

# Return results for all accounts
count(listAccountCodes(), "published inthelast month")
ac <- account("QUIR01BA", key="<my key>")

# A single number counting the mentions published in the last week.
count(ac, "published inthelast week")

# The number of relevant mentions published in the last month
count(ac, "published inthelast month and relevancy isnt irrelevant")

# As above, but grouped by publication day
count(ac, "published inthelast month and relevancy isnt irrelevant",
      groupby="published")
}

