% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{ensureAuthenticated}
\alias{ensureAuthenticated}
\title{Check that the user is authenticated.}
\usage{
ensureAuthenticated(authentication)
}
\arguments{
\item{authentication}{The authentication object}
}
\description{
Check that the user is authenticated, and provide a consistent error message.
}
\author{
Constance Neeser
}
