% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accounts.R
\name{list_accounts}
\alias{list_account_codes}
\alias{list_accounts}
\title{List accounts you have access to}
\usage{
list_accounts(auth = pkg.env$defaultAuthentication, key = NULL,
  user = NULL, password = NULL)

list_account_codes(auth = pkg.env$defaultAuthentication, key = NULL,
  user = NULL, password = NULL)
}
\description{
This returns a data frame listing the accounts that you have access to,
along with their name and status.
}
\section{Functions}{
\itemize{
\item \code{list_account_codes}: Returns a character vector of account codes that you have access to.
}}
\examples{
\dontrun{

list_accounts(key = "my api key")

list_accounts(user = "rudy.neeser@brandseye.com",
              password = "my brandseye password")

auth <- authentication(user = "rudy.neeser@brandseye.com",
                       password = "my brandseye password")
list_accounts(auth)

}
\dontrun{
# Get the number of mentions published in the last day across all
# of your accounts.
account_count(list_account_codes(), "published inthelast day")
}
}
\author{
Rudy Neeser
}

