% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{listAccounts}
\alias{listAccountCodes}
\alias{listAccounts}
\title{List accounts you have access to}
\usage{
listAccounts(auth = defaultAuthentication, key = NULL, user = NULL,
  password = NULL)

listAccountCodes(auth = defaultAuthentication, key = NULL, user = NULL,
  password = NULL)
}
\description{
This returns a data frame listing the accounts that you have access to,
along with their name and status.
}
\section{Functions}{
\itemize{
\item \code{listAccountCodes}: Returns a character vector of account codes that you have access to.
}}
\examples{
listAccounts(key = "my api key")

listAccounts(user = "rudy.neeser@brandseye.com",
             password = "my brandseye password")

auth <- authentication(user = "rudy.neeser@brandseye.com",
                       password = "my brandseye password")
listAccounts(auth)
# Get the number of mentions published in the last day across all
# of your accounts.
count(listAccountCodes(), "published inthelast day")
}

