% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{print.brandseye.clientService}
\alias{print.brandseye.clientService}
\title{Prints a client service S3 class.}
\usage{
\method{print}{brandseye.clientService}(client_service)
}
\description{
Prints a client service S3 class.
}

