% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mentions.R
\name{print.mention.results}
\alias{print.mention.results}
\title{Prints a summary of the results from a call to \code{\link{account_mentions}}}
\usage{
\method{print}{mention.results}(mentions, ...)
}
\description{
Prints a summary of the results from a call to \code{\link{account_mentions}}
}
\author{
Rudy Neeser
}

