% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{print.summary.brandseye.account}
\alias{print.summary.brandseye.account}
\title{Prints a summary of a BrandsEye \code{\link{account}}}
\usage{
\method{print}{summary.brandseye.account}(s.ac, ...)
}
\description{
Prints a summary of a BrandsEye \code{\link{account}}
}
\author{
Rudy Neeser
}

