% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\title{Create sentiment factors}
\usage{
sentiment(value)
}
\arguments{
\item{value}{The integer value of the sentiment you want to represent.}
}
\description{
Create representations of sentiment as ordered factors.
}
\details{
BrandsEye represents sentiment as ordered factors between -5 and 5. 
0 is never used, but is represented here to allow for easy querying.

Sentiment is represents as ordered factors with the class 
\code{c("sentiment", "ordered", "factor")}.
}
\examples{
 

sentiment(1) < sentiment(2)

\dontrun{

library(dplyr)
account_count("QUIR01BA", "published inthelast month", groupby="sentiment") \%>\%
  filter(sentiment <= sentiment(-1))
}


}
\author{
Constance Neeser
}
