% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{summary.brandseye.account}
\alias{summary.brandseye.account}
\title{Summarising accounts}
\usage{
\method{summary}{brandseye.account}(account, ...)
}
\description{
\code{summary} method for class "\code{brandseye.account}".
}

