% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account.brands}
\alias{account.brands}
\alias{account.brands.brandseye.account}
\alias{account.brands.character}
\title{Listing brands in an account}
\usage{
account.brands(account, ...)

\method{account.brands}{brandseye.account}(account, .process = TRUE)

\method{account.brands}{character}(account)
}
\description{
This returns a \code{data.frame} containing brand IDs, their names,
whether they've been deleted or not, and the ID of the parent brand.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: 

\item \code{character}: 
}}
\examples{
\dontrun{
# Fetching brands for a particular account
account.brands("QUIR01BA")
}
}

