% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account}
\alias{account}
\title{Provides access to a BrandsEye account}
\usage{
account(code, key = NULL, user = NULL, password = NULL,
  auth = pkg.env$defaultAuthentication)
}
\description{
Provides access to a BrandsEye account
}
\details{
Creates an object representing a BrandsEye account. It can be used to easily perform
various queries on the account.

\code{summary} can be used to briefly summarise an account.
}
\examples{
\dontrun{
ac <- account("QUIR01BA", user = "rudy.neeser@brandseye.com",
              password = "This is not my real password")
account("QUIR01BA", key="<api key here>")

# Have a brief summary of an account
summary(ac)
}
}
\seealso{
\code{\link{listAccounts}} to see the accounts that you have access to.

\code{\link{listAccountCodes}} for a vector of account codes that you have

\code{\link{account.brands}} for listing the brands in an account.

\code{\link{account.phrases}} for listing the phrases used in an account.

\code{\link{account_tags}} for listing the tags used in an account.
}

