% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account}
\alias{account}
\title{Provides access to a BrandsEye account}
\usage{
account(code, auth = pkg.env$defaultAuthentication, key = NULL,
  user = NULL, password = NULL)
}
\description{
Provides access to a BrandsEye account
}
\details{
Creates an object representing a BrandsEye account. It can be used to easily perform
various queries on the account.
}
\examples{
ac <- account("QUIR01BA", user = "rudy.neeser@brandseye.com",
              password = "This is not my real password")
account("QUIR01BA", key="<api key here>")
}

