% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brandseyer.R
\name{account}
\alias{account}
\title{Provides access to a BrandsEye account}
\usage{
account(code, key = NULL, user = NULL, password = NULL)
}
\description{
Provides access to a BrandsEye account
}
\examples{
ac <- account("QUIR01BA", user = "rudy.neeser@brandseye.com",
              password = "This is not my real password")
account("QUIR01BA", key="<api key here>")
}

