% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_brands}
\alias{account_brands}
\alias{account_brands.brandseye.account}
\alias{account_brands.character}
\alias{account_brands.factor}
\alias{account_brands.list}
\title{Listing brands in an account}
\usage{
account_brands(account, ...)

\method{account_brands}{brandseye.account}(account, .process = TRUE)

\method{account_brands}{list}(accounts)

\method{account_brands}{character}(account)

\method{account_brands}{factor}(account)
}
\description{
This returns a \code{data.frame} containing brand IDs, their names,
whether they've been deleted or not, and the ID of the parent brand.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: Returns brand information for a particular \code{\link{account}} object.

\item \code{list}: Returns brand information for a list of accounts

\item \code{character}: Returns brand information for an account identified by an account code.

\item \code{factor}: Returns brand information for an account identified by a factor of its account code.
}}
\examples{
\dontrun{
# Fetch brands for an \\code{\\link{account}} object.
ac <- account("QUIR01BA")
account_brands(ac)
}
\dontrun{
# Returns a data.frame containing brand information for two accounts
account_brands(c("QUIR01BA", "BEAD33AA"))

# Return brand information for all accounts that you have access to
account_brands(list_account_codes())
}
\dontrun{
# Fetching brands for a particular account
account_brands("QUIR01BA")
}

}
\author{
Rudy Neeser
}

