% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_code}
\alias{account_code}
\alias{account_code.brandseye.account}
\alias{account_code.list}
\title{Returns a BrandsEye \code{\link{account}}'s code.}
\usage{
account_code(account)

\method{account_code}{brandseye.account}(account)

\method{account_code}{list}(accounts)
}
\description{
Returns a BrandsEye \code{\link{account}}'s code.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: 

\item \code{list}: 
}}
\author{
Rudy Neeser
}

