% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mentions.R
\name{account_mentions}
\alias{account_mentions}
\alias{account_mentions.brandseye.account}
\alias{account_mentions.character}
\alias{account_mentions.factor}
\title{Read mentions from your account}
\usage{
account_mentions(account, ...)

\method{account_mentions}{character}(code, filter, limit = 30, offset = 0,
  include, authentication = pkg.env$defaultAuthentication, all = FALSE,
  showProgress = length(code) != 0)

\method{account_mentions}{brandseye.account}(account, ...)

\method{account_mentions}{factor}(account, ...)
}
\arguments{
\item{limit}{The maximum number of mentions to be returned}

\item{offset}{Mentions are returned in an order. Offset says how many of the
first mentions should be skipped.}

\item{include}{A character vector of extra information to include in the mentions
You can see this list in the api documentation available at
\url{https://api.brandseye.com/docs}}

\item{all}{Set to true if you would like to return all mentions from the account.}
}
\value{
Returns an object of class \code{mention.results}, which is a list
containing at least the following items:
\itemize{
\item \code{mentions}, a \code{data.frame} of the mention data.
\item \code{media}, a \code{data.frame} listing mime types and urls for media
      associated with each mention.
\item \code{tags}, a \code{data.frame} containing the tags that match the mentions.
\item \code{phrases}, a \code{data.frame} listing the phrases that a mention matched.
\item \code{sentiment}, a \code{data.frame} listing the sentiment associated with a mention, possibly
      more than one per mention.
}
}
\description{
Read mentions from your account
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{brandseye.account}: 

\item \code{factor}: For querying mentions from accounts encoded as \code{factor}s.
}}
\examples{
\dontrun{

# Read mentions using the default authentication
account_mentions("QUIR01BA", "published inthelast day")

}
}
\author{
Rudy Neeser
}
\seealso{
\code{\link{account_phrases}} to see a list of phrases in an account, with
         their associated brands.

\code{\link{account_brands}} for a list of brands associated with the account.
}

