% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{account_name}
\alias{account_name}
\alias{account_name.brandseye.account}
\alias{account_name.character}
\alias{account_name.factor}
\title{Returns the human readable name of an account}
\usage{
account_name(account)

\method{account_name}{brandseye.account}(account)

\method{account_name}{character}(accounts)

\method{account_name}{factor}(accounts)
}
\description{
Returns the human readable name of an account
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseye.account}: Returns the name from an \code{\link{account}} object.

\item \code{character}: Returns the name from a vector of account codes.

\item \code{factor}: Returns the name of an account given by a factor of the account code.
}}

\author{
Constance Neeser
}
