% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{account_update}
\alias{account_update}
\alias{account_update.character}
\title{Updates mentions for accounts}
\usage{
account_update(account, ...)

\method{account_update}{character}(code, filter, tag, sentiment, media,
  auto.confirm = FALSE, authentication = pkg.env$defaultAuthentication)
}
\arguments{
\item{code}{A character string for the account to update}

\item{filter}{A filter selecting mentions to update}

\item{tag}{Either an id of a tag to update the selected mentions with, or
a character string holding a tag name.}

\item{sentiment}{An integer representing sentiment. A value between -5 and 5. 
0 is not a valid number, and 1 represents neutral sentiment.}

\item{auto.confirm}{Some actions, such as setting sentiment, will ask for interactive
confirmation of the action. If you're running from a script, this 
can be inconvenient. You can override that behaviour by setting
this parameter to \code{TRUE}. Buyer beware.}
}
\value{
Nothing useful yet. This method is a placeholder for testing, and 
its return values will likely change. Its interface will probably remain stable,
although extra arguments may be added.
}
\description{
Updates mentions for accounts

Updates mentions for accounts supplied as a vector of character codes.
}
\details{
This function is solely for testing purposes, and its behaviour
may change in future versions.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}

\author{
Constance Neeser

Lara Basson
}
