% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brandseyer.R
\docType{package}
\name{brandseyer}
\alias{BrandsEye}
\alias{BrandsEyeR}
\alias{brandseye}
\alias{brandseyer}
\alias{brandseyer-package}
\title{brandseyer: access and perform analytics on your BrandsEye accounts}
\description{
BrandsEye is a company that provides online monitoring, analytics,
and insights for social media data. We provide an extensive JSON / RESTful API
to access your data. This library is a wrapper for that data, providing
easy access to the data in your various BrandsEye accounts for use in
any R analyses and data visualisations.
}
\details{
The home page is \url{https://github.com/brandseye/brandseyer/}

If you have found a bug in the library, you can report it using the
library's GitHub issue tracker: \url{https://github.com/brandseye/brandseyer/issues}
}

