% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{print.brandseye.account}
\alias{print.brandseye.account}
\title{Print's details of a brandseye account.}
\usage{
\method{print}{brandseye.account}(account, ...)
}
\description{
Print's details of a brandseye account.
}
\author{
Constance Neeser
}
